requirements.txt
python-telegram-bot==21.4
PyYAML==6.0.2
pytest==8.3.2
